backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pandas==2.2.1
